fix multi-app environments
multi-app environments had a number of fundamental problems.

* the method for merging individual application's environments was
  extremely simplistic and could result in parts of an app's
  environment being reset.

* application environments were not necessarily based on the current
  environment and when applied could reset the user's environment.
  This problem was found for environments created by shell scripts
  where the user's profile or .cshrc was sourced as part of the shell
  invocation, in many cases resulting in a reset of the environment to
  its inital state at login.

To avoid these problems, a new restriction was placed upon the
application specific modules.  They must create their environments
based upon the current one, merging their changes into it.  When
creating a merged environment, the application modules are
successively invoked on a copy of the current environment and the
final environment is used.  Previously the modules' environments were
"merged" in a rather inflexible manner, resulting in overwriting
changes.  This also means that 1) cached application environments
can't be used during a merger and 2) application environments can't be
cached during the course of a merger.