merged env()/str() Include behavior into list of vars; str now takes variable names, normalized return behavior of env() in list/scalar context.
env() & str() variable name, Include & Exclude behavior was too
complex.  the Include option was partially duplicating the list of
names at the front of the argument list.  The Include option was
removed and the list of variable names was upgraded to a list of match
specifications.  str() now has the same argument signature as env().

The return from env() in different contexts was too inflexible (and
maybe a bit inconsistent).  It is now deterministic and documented
(and perhaps too flexible).