use Object::ID to track parent object for cached _app objects
When an App::Env object is cached, it's really the underlying _app
object that is cached.  The cached object keeps track of its "parent"
App::Env object to ensure that only that object can uncache it.

Previously refaddr was used to identify the parent object; this may
not stay constant.  Now Object::ID is used.