explicitly ignore prototypes on Capture::Tiny::capture.
Capture::Tiny::capture is prototyped as (&;@).  It is called with the
correct argument type (a leading coderef).  However, becaue of the
prototype, which requires an explicit block or sub {}, calling it as

  capture( $sub );

will elicit a compile time error.

App::Env lazy-loads Capture::Tiny and thus nominally avoids the
prototype check.  However, if Capture::Tiny is explicitly loaded prior
to App::Env, the prototype check will be performed when App::Env is
compiled.

To avoid the action at a distance this creates, calls to capture
now explicitly avoid prototype checking.